+++ src/onegov.town/onegov/town/forms/reservation.py
@@ -1,3 +1,6 @@
+import sedate
+
+from datetime import time
 from onegov.form import Form
 from onegov.town import _
 from wtforms.fields import TextField
@@ -7,6 +10,31 @@ from wtforms.validators import Email, InputRequired, Regexp
 
 class ReservationForm(Form):
 
+    def apply_model(self, reservation):
+
+        self.email.data = reservation.email
+
+        if hasattr(self, 'start'):
+            self.start.data = '{:%H:%M}'.format(reservation.display_start())
+
+        if hasattr(self, 'end'):
+            self.end.data = '{:%H:%M}'.format(reservation.display_end())
+
+        if hasattr(self, 'quota'):
+            self.quota.data = reservation.quota
+
+    def get_date_range(self):
+        if self.allocation.partly_available:
+            return sedate.get_date_range(
+                sedate.to_timezone(
+                    self.allocation.start, self.allocation.timezone
+                ),
+                time(*(int(p) for p in self.data['start'].split(':'))),
+                time(*(int(p) for p in self.data['end'].split(':')))
+            )
+        else:
+            return self.allocation.start, self.allocation.end
+
     @classmethod
     def for_allocation(cls, allocation):
 
@@ -14,6 +42,7 @@ class ReservationForm(Form):
             pass
 
         form_class = AdaptedReservationForm
+        form_class.allocation = allocation
 
         if allocation.partly_available:
             time_validator = Regexp(r'[0-9]{2}:[0-9]{2}', message=_(
@@ -33,7 +62,7 @@ class ReservationForm(Form):
                 default='{:%H:%M}'.format(allocation.display_end())
             )
 
-        form_class.e_mail = EmailField(
+        form_class.email = EmailField(
             label=_("E-Mail"),
             validators=[InputRequired(), Email()]
         )