+++ src/onegov.town/HISTORY.rst
@@ -4,6 +4,13 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Adds an extra confirmation step to the reservations to be more consistent
+  with the way form and event submissions work.
+
+- Adds the ability to reserve parts of an allocation. Allocations in rooms are
+  partly reservable by default.
+  [href]
+
 0.9.2 (2015-11-24)
 ~~~~~~~~~~~~~~~~~~~